reduce size of layer id

+++ rgis-layer-id/src/lib.rs
@@ -1,4 +1,4 @@
 #[derive(
     Copy, Clone, Debug, Eq, Ord, PartialEq, PartialOrd, Hash, bevy::ecs::component::Component,
 )]
-pub struct LayerId(pub i64); // todo: we can make this an i16
+pub struct LayerId(pub u16);
